app/requirements.txt
bokeh==2.4.3
joblib==1.1.1
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
PyYAML==6.0.1
requests==2.28.1
scikit_learn==1.3.1
streamlit==1.24.0
tensorflow
